refactor: move ``testDBConnection` to `utils` folder
